refactor(imu): address CodeRabbit round 7 findings
- main.c: remove redundant (uint8_t) casts on imu_i2c_addr_t enum constants
  (uint8_t underlying type satisfies uint8_t parameter directly)
- imu_task.h: add k_imu_task_period_ticks = 5U to imu_task_timing_t;
  update @invariant to reference enum constant; update docblock details
- imu_task.c: remove local k_imu_task_period_ticks from imu_task_cfg_t
  (now defined in imu_task.h); change while(true) to while(1)
- rx_bmp280_regs.h: rename k_bmp280_shift_lsb/xlsb to
  k_bmp280_shift_lsb_left/xlsb_right to clarify shift direction;
  split bmp280_cfg_t into three focused enums (bmp280_ctrl_meas_t,
  bmp280_config_t, bmp280_status_mask_t); add @invariant to
  bmp280_write_size_t docblock
- rx_bmp280.c: update shift constant references; add @since tag to file
  docblock; fix assert cast from (temp_bit_constants_t) to (int32_t);
  drop unused local typedef name; extract internal_read_and_compensate_adc()
  helper from rx_bmp280_read() with full Doxygen docblock
- rx_bmp280.h: update shift constant names in Doxygen comment
- rx_bno055_regs.h: add U suffix to bno055_read_size_t enum values
- rx_bno055.c: fix internal_init_enter_ndof assert from s_tag to s_bus_name
- test_rx_bmp280.c: replace manual calib setup in zero-var1 test with
  internal_load_invalid_calib_p1_zero() helper call

+++ star-rx72n-firmware/src/main.c
@@ -1721,7 +1721,7 @@ static void internal_register_system_buses(void)
   err = rx_bus_config_init_i2c(&s_i2c1_imu_config,
                                "i2c1",                                       /* name: matches rx_bno055.c s_bus_name */
                                k_riic_channel_1,                             /* channel: RIIC1 */
-                               (uint8_t)k_i2c_addr_bno055,                  /* device_addr: BNO055 (COM3/ADR=LOW) */
+                               k_i2c_addr_bno055,                            /* device_addr: BNO055 (COM3/ADR=LOW) */
                                k_rx_p2_0,                                    /* sda_pin: P2.0 = SDA1 */
                                k_rx_p2_1,                                    /* scl_pin: P2.1 = SCL1 */
                                k_i2c_frequency_400khz);                      /* frequency_hz: 400 kHz fast mode */
@@ -1733,7 +1733,7 @@ static void internal_register_system_buses(void)
   err = rx_bus_config_init_i2c(&s_i2c1_baro_config,
                                "i2c1_baro",                                  /* name: matches rx_bmp280.c s_bus_name */
                                k_riic_channel_1,                             /* channel: RIIC1 (same as i2c1) */
-                               (uint8_t)k_i2c_addr_bmp280,                  /* device_addr: BMP280 (SDO=LOW) */
+                               k_i2c_addr_bmp280,                            /* device_addr: BMP280 (SDO=LOW) */
                                k_rx_p2_0,                                    /* sda_pin: P2.0 = SDA1 */
                                k_rx_p2_1,                                    /* scl_pin: P2.1 = SCL1 */
                                k_i2c_frequency_400khz);                      /* frequency_hz: 400 kHz fast mode */
